VULCAN-10 Move each test to its module
