+++ modules/apps/foundation/vulcan/vulcan-message-hal/build.gradle
@@ -6,4 +6,7 @@ dependencies {
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	provided project(":apps:foundation:vulcan:vulcan-api")
 	provided project(":apps:foundation:vulcan:vulcan-wiring-osgi")
+
+	testCompile group: "com.google.code.gson", name: "gson", version: "2.8.1"
+	testCompile project(":apps:foundation:vulcan:vulcan-jaxrs-json")
 }
