+++ modules/apps/foundation/vulcan/vulcan-message-json-ld/build.gradle
@@ -5,4 +5,7 @@ dependencies {
 	provided group: "javax.ws.rs", name: "javax.ws.rs-api", version: "2.0.1"
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	provided project(":apps:foundation:vulcan:vulcan-api")
+
+	testCompile group: "com.google.code.gson", name: "gson", version: "2.8.1"
+	testCompile project(":apps:foundation:vulcan:vulcan-jaxrs-json")
 }
