+++ modules/apps/foundation/vulcan/vulcan-sample-liferay-portal/build.gradle
@@ -10,4 +10,6 @@ dependencies {
 	provided project(":apps:foundation:vulcan:vulcan-api")
 	provided project(":apps:foundation:vulcan:vulcan-liferay-portal")
 	provided project(":apps:foundation:vulcan:vulcan-wiring-osgi")
+
+	testCompile group: "org.mockito", name: "mockito-all", version: "1.8.4"
 }
