--- modules/apps/foundation/vulcan/vulcan-test/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay Vulcan Test
-Bundle-SymbolicName: com.liferay.vulcan.test
-Bundle-Version: 1.0.0
